app/requirements.txt
Flask
pandas
matplotlib
pyyaml